feat: 최신 Layout Componet를 사용하는 페이지로 교체
