+++ pages/index.tsx
@@ -1,59 +1,68 @@
-import Layout from '@/components/common/Layout';
-import ExchangeRateBox from '@/components/dividend/ExchangeRate';
 import ModeController from '@/components/ModeController';
-import CommonCenter from '@/components/common/Center';
-import CommonHeader from '@/components/common/Header';
-import DetailStocksSection from '@/components/portfolio/portfolioSections/DetailStocksSection';
-import TotalStatisticsSection from '@/components/portfolio/portfolioSections/TotalStatisticsSection';
-import { useMyPortFolio } from '@/hook/useMyPortFolio';
-import NotifyMessageSection from '@/components/portfolio/portfolioSections/NotifyMessageSection';
 import CommonBar from '@/components/common/Bar';
-import SimpleDividentScheduleSection from '@/components/portfolio/portfolioSections/SimpleDividentScheduleSection';
+import CommonHeader from '@/components/common/Header';
+import CenterContent from '@/components/commonV2/CenterContent';
+import LayoutV2 from '@/components/commonV2/Layout';
+import Padding from '@/components/commonV2/Padding';
+import ExchangeRateBox from '@/components/dividend/ExchangeRate';
 import ChartSection from '@/components/portfolio/portfolioSections/ChartSection';
+import DetailStocksSection from '@/components/portfolio/portfolioSections/DetailStocksSection';
+import NotifyMessageSection from '@/components/portfolio/portfolioSections/NotifyMessageSection';
+import SimpleDividentScheduleSection from '@/components/portfolio/portfolioSections/SimpleDividentScheduleSection';
+import TotalStatisticsSection from '@/components/portfolio/portfolioSections/TotalStatisticsSection';
+import { useAnnualDividend } from '@/hook/useAnnualDividend';
+import { useMyPortFolio } from '@/hook/useMyPortFolio';
+import { useMonthlyCalanderDividendQuery } from '@/hook/useQueryHook/useMonthlyCalanderDividendQuery';
 import { CircularProgress } from '@mui/material';
 
-import Head from 'next/head';
+export default function V2Page() {
+  const { status } = useMyPortFolio();
+  useAnnualDividend();
+  useMonthlyCalanderDividendQuery();
 
-const MainPage = () => {
-  const { myPortFolioData, status } = useMyPortFolio();
-
-  if (status === 'loading') {
-    return (
-      <Layout>
-        <CommonCenter>
+  return (
+    <LayoutV2
+      showBottomNavigator
+      headMetaProps={{
+        title: '스노우볼 - 배당 모아보기',
+        image: '/icon/snow_logo.png',
+      }}
+    >
+      {status === 'loading' ? (
+        <CenterContent>
           <CircularProgress />
-        </CommonCenter>
-      </Layout>
-    );
-  }
+        </CenterContent>
+      ) : (
+        <div>
+          <Padding paddingLeft={18} paddingRight={18} paddingTop={21}>
+            <CommonHeader>
+              <ModeController />
+            </CommonHeader>
+          </Padding>
 
-  const hasNotAssets = myPortFolioData?.assetDetails.length;
+          <Padding paddingLeft={18} paddingRight={18}>
+            <ExchangeRateBox />
+          </Padding>
+          <Padding paddingLeft={18} paddingRight={18}>
+            <ChartSection />
+          </Padding>
+          <Padding paddingLeft={18} paddingRight={18}>
+            <TotalStatisticsSection />
+          </Padding>
+          <Padding paddingLeft={18} paddingRight={18}>
+            <SimpleDividentScheduleSection />
+          </Padding>
 
-  if (status === 'success' && hasNotAssets) {
-    return (
-      <>
-        <Head>
-          <title>나의 포트폴리오 - 스노우볼</title>
-          <meta name="description" content="Generated by create next app" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-        <Layout>
-          <CommonHeader>
-            <ModeController />
-          </CommonHeader>
-          <ExchangeRateBox />
-          <ChartSection />
-          <TotalStatisticsSection />
-          <SimpleDividentScheduleSection />
           <CommonBar />
-          <DetailStocksSection />
-          <NotifyMessageSection />
-        </Layout>
-      </>
-    );
-  }
 
-  return null;
-};
-
-export default MainPage;
+          <Padding paddingLeft={18} paddingRight={18}>
+            <DetailStocksSection />
+          </Padding>
+          <Padding paddingLeft={18} paddingRight={18}>
+            <NotifyMessageSection />
+          </Padding>
+        </div>
+      )}
+    </LayoutV2>
+  );
+}
